pyproject.toml
[project]
name = "backend-in"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "livekit-agents[google,openai,silero,turn-detector] (>=1.0,<2.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "livekit-plugins-noise-cancellation (>=0.2,<1.0)",
    "fastapi (>=0.115.14,<0.116.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "email-validator (>=2.2.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false